requirements.txt




Flask==2.1.0
gcloud==0.18.3
glob2==0.7
google-api-core==2.11.0
google-auth==2.15.0
google-cloud==0.34.0
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
gunicorn==20.1.0
httplib2==0.21.0
oauth2client==4.1.3
panda==0.3.1
pandas==1.5.2
pandas-datareader==0.10.0
pandas-io==0.0.1
pathlib==1.0.1
pip==22.2.2
prometheus-client==0.14.1
py4j==0.10.9.5
pyarrow==10.0.1
pyspark==3.3.1
python-dateutil==2.8.2
requests==2.28.1
requests-file==1.5.1
six==1.16.0
urllib3==1.26.11
